Make context optional in reading/downloading
