+++ tox.ini
@@ -65,6 +65,8 @@ commands =
 [testenv:integration]
 deps =
     {[base]testdeps}
+passenv =
+    ACCESS_TOKEN
 setenv =
     DB_API_URL = {env:DB_API_URL:http://localhost:8000}
 commands =
